# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Nexura API
  version: 1.0.0
  description: Nexura Phase 1 core foundation API.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1
security:
  - SessionCookie: []
paths:
...
  /guilds:
    get:
      operationId: listGuilds
      summary: List Discord guilds visible to the authenticated user
      responses:
        '200':
          description: Guild list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Guild'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
  /guilds/{guildId}:
    get:
      operationId: getGuild
      summary: Get a manageable guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{17,20}$'
      responses:
        '200':
          description: Guild details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guild'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: nexura.sid
  schemas:
...
    Guild:
      type: object
      required: [id, name, icon, botPresent, canManage, permissionRole]
      properties:
        id: { type: string, pattern: '^\d{17,20}$' }
        name: { type: string }
        icon: { type: [string, 'null'] }
        botPresent: { type: boolean }
        canManage: { type: boolean }
        permissionRole:
          type: [string, 'null']
          enum: [OWNER, ADMINISTRATOR, MANAGER, null]
    Problem:
      type: object
      required: [type, title, status, detail, instance, requestId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        requestId: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Authentication is required or invalid
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: The authenticated user cannot access the resource
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    BadGateway:
      description: Discord API is unavailable
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }